requirements.txt
streamlit
google-genai
python-dotenv
youtube_transcript_api